Extract csv generator to a separate package
The same code was being used in two places to generate and return csv
files. This extracts it to one place.
